requirements.txt
apache-airflow==2.9.1
pandas
requests
apache-airflow-providers-cncf-kubernetes==9.0.0